feat: finish slicing story page
